Move InvokeAsync methods to be extensions on IRenderedFragmentBase
