+++ src/bunit.core/ComponentParameterCollectionBuilder.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Threading.Tasks;
+using Bunit.Extensions;
 using Microsoft.AspNetCore.Components;
 
 namespace Bunit
