+++ src/bunit.core/ComponentParameterFactory.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Threading.Tasks;
+using Bunit.Extensions;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Rendering;
