+++ src/bunit.core/Extensions/WaitForHelpers/WaitForHelper.cs
@@ -53,7 +53,6 @@ namespace Bunit.Extensions.WaitForHelpers
 			completionSouce = new TaskCompletionSource<object?>();
 			timer = new Timer(OnTimeout, this, Timeout.InfiniteTimeSpan, Timeout.InfiniteTimeSpan);
 
-			OnAfterRender(this, EventArgs.Empty);
 			this.renderedFragment.OnAfterRender += OnAfterRender;
 			OnAfterRender(this, EventArgs.Empty);
 			StartTimer(timeout);
